requirements.txt
pyyaml>=6.0
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
jinja2>=3.1.0
rich>=13.0.0

# For human review interface
streamlit>=1.29.0
pandas>=2.0.0
plotly>=5.18.0

# Optional: MongoDB storage
# pymongo>=4.6.0
